+++ tests/test_service_providers.py
@@ -3,6 +3,7 @@ from sqlalchemy.orm import Session, sessionmaker
 from limberframework.config.config_service_provider import ConfigServiceProvider
 from limberframework.database.connections import SqliteConnection
 from limberframework.database.database_service_provider import DatabaseServiceProvider
+from limberframework.authentication.authentication_service_provider import AuthServiceProvider
 
 def test_database_service_provider():
     mock_app = MagicMock()
@@ -15,5 +16,12 @@ def test_config_service_provider():
     mock_app = MagicMock()
     config = ConfigServiceProvider(mock_app)
     config.register()
+
+    assert mock_app.bind.called_once()
+
+def test_authentication_service_provider():
+    mock_app = MagicMock()
+    auth = AuthServiceProvider(mock_app)
+    auth.register()
 
     assert mock_app.bind.called_once()