+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Soft delete property to Model class to remove Models without deleting from the database.
 - Tests for config/config module.
 - Test for creating a PostgresConnection instance.
+- Authentication service provider along with HTTP Basic and API Key authenticators.
 
 ### Fixed
 - Resolve SQLite single thread warning by turning off 'same thread check'.